+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.20 branch
+### 2.20.1
+* added `IntegrationTestCase` and `ValidationTestCase` classes.
+
 ### 2.20.0
 * added initial schema of the plugin database. Removed `BaseUpdateConsole` and
     `UpdateShell` classes;